requirements.txt
# requirements.txt im Hauptverzeichnis überprüfen/erstellen
cat > requirements.txt << 'EOF'
# Core
SQLAlchemy>=2.0.0
alembic>=1.12.0
requests>=2.31.0
python-dotenv>=1.0.0

# Optional but helpful
pandas>=2.0.0
numpy>=1.24.0

# Testing
pytest>=7.0.0

# Development
ipython>=8.0.0
EOF

# Pakete installieren
pip install -r requirements.txt